+++ src/Plugins/SmartStore.FacebookAuth/Core/FacebookProviderAuthorizer.cs
@@ -2,19 +2,22 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Web;
 using System.Web.Mvc;
 using DotNetOpenAuth.AspNet;
 using DotNetOpenAuth.AspNet.Clients;
+using Newtonsoft.Json.Linq;
 using SmartStore.Core.Domain.Customers;
 using SmartStore.Services;
 using SmartStore.Services.Authentication.External;
 
 namespace SmartStore.FacebookAuth.Core
 {
-    public class FacebookProviderAuthorizer : IOAuthProviderFacebookAuthorizer
+	public class FacebookProviderAuthorizer : IOAuthProviderFacebookAuthorizer
 	{
 		#region Fields
 		
@@ -95,13 +98,39 @@ namespace SmartStore.FacebookAuth.Core
             return state;
         }
 
+		private string GetEmailFromFacebook(string accessToken)
+		{
+			var result = "";
+			var webRequest = WebRequest.Create("https://graph.facebook.com/me?fields=email&access_token=" + EscapeUriDataStringRfc3986(accessToken));
+
+			using (var webResponse = webRequest.GetResponse())
+			using (var stream = webResponse.GetResponseStream())
+			using (var reader = new StreamReader(stream))
+			{
+				var strResponse = reader.ReadToEnd();
+				var info = JObject.Parse(strResponse);
+
+				if (info["email"] != null)
+				{
+					result = info["email"].ToString();
+				}
+			}
+			return result;
+		}
+
 		private void ParseClaims(AuthenticationResult authenticationResult, OAuthAuthenticationParameters parameters)
         {
 			var claims = new UserClaims();
 			claims.Contact = new ContactClaims();
-			
+
 			if (authenticationResult.ExtraData.ContainsKey("username"))
+			{
 				claims.Contact.Email = authenticationResult.ExtraData["username"];
+			}
+			else
+			{
+				claims.Contact.Email = GetEmailFromFacebook(authenticationResult.ExtraData["accesstoken"]);
+			}
 
 			claims.Name = new NameClaims();